python/requirements.txt
ecdsa==0.13
python-jose==0.6.1
